apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.12
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
httpx==0.28.1
pandas==2.3.2
openai==1.108.1
numpy==2.3.3
python-dateutil==2.9.0.post0
pgvector==0.4.1
pytest==8.4.2
email-validator==2.3.0
PyMySQL==1.1.2
simple-salesforce==1.12.9
bcrypt==4.1.3

